Fix serialization issues with the AWS Console test requests.
Test requests serialize null headers and queries, instead
of omitting the fields. These changes make the deserialization
work by creating default structs, so we don't have to add more
optionals.
