+++ aws_lambda_events/Cargo.toml
@@ -26,7 +26,7 @@ serde_derive = "^1"
 serde_json = "^1"
 bytes = { version = "1", features = ["serde"] }
 chrono = { version = "^0.4.4", features = ["serde"] }
-query_map = { version = "^0.3", features = ["serde"] }
+query_map = { version = "^0.4", features = ["serde"] }
 
 [dev-dependencies]
 pretty_assertions = "0.7"
