+++ aws_lambda_events/src/custom_serde/mod.rs
@@ -1,11 +1,16 @@
 use base64::{decode, encode};
 use chrono::{DateTime, Duration, TimeZone, Utc};
-use http::{HeaderMap, Method};
 use serde;
-use serde::de::{Deserialize, Deserializer, Error as DeError, Unexpected, Visitor};
-use serde::ser::{Error as SerError, SerializeMap, Serializer};
+use serde::de::{Deserialize, Deserializer, Error as DeError};
+use serde::ser::Serializer;
 use std::collections::HashMap;
-use std::fmt;
+
+pub(crate) mod codebuild_time;
+
+mod headers;
+pub(crate) use self::headers::*;
+
+pub(crate) mod http_method;
 
 fn normalize_timestamp<'de, D>(deserializer: D) -> Result<(u64, u64), D::Error>
 where
@@ -179,161 +184,6 @@ where
 {
     let minutes = f64::deserialize(deserializer)?;
     Ok(Duration::minutes(minutes as i64))
-}
-
-/// Serialize a http::HeaderMap into a serde str => Vec<str> map
-pub(crate) fn serialize_multi_value_headers<S>(
-    headers: &HeaderMap,
-    serializer: S,
-) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    let mut map = serializer.serialize_map(Some(headers.keys_len()))?;
-    for key in headers.keys() {
-        let mut map_values = Vec::new();
-        for value in headers.get_all(key) {
-            map_values.push(value.to_str().map_err(S::Error::custom)?)
-        }
-        map.serialize_entry(key.as_str(), &map_values)?;
-    }
-    map.end()
-}
-
-/// Serialize a http::HeaderMap into a serde str => str map
-pub(crate) fn serialize_headers<S>(headers: &HeaderMap, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    let mut map = serializer.serialize_map(Some(headers.keys_len()))?;
-    for key in headers.keys() {
-        let map_value = headers[key].to_str().map_err(S::Error::custom)?;
-        map.serialize_entry(key.as_str(), map_value)?;
-    }
-    map.end()
-}
-
-pub mod http_method {
-    use super::*;
-
-    pub fn serialize<S: Serializer>(method: &Method, ser: S) -> Result<S::Ok, S::Error> {
-        ser.serialize_str(method.as_str())
-    }
-
-    struct MethodVisitor;
-    impl<'de> Visitor<'de> for MethodVisitor {
-        type Value = Method;
-
-        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-            write!(formatter, "valid method name")
-        }
-
-        fn visit_str<E: DeError>(self, val: &str) -> Result<Self::Value, E> {
-            if val.is_empty() {
-                Ok(Method::GET)
-            } else {
-                val.parse()
-                    .map_err(|_| DeError::invalid_value(Unexpected::Str(val), &self))
-            }
-        }
-    }
-
-    pub fn deserialize<'de, D>(de: D) -> Result<Method, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        de.deserialize_str(MethodVisitor)
-    }
-
-    pub fn deserialize_optional<'de, D>(deserializer: D) -> Result<Option<Method>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s: Option<String> = Option::deserialize(deserializer)?;
-        if let Some(val) = s {
-            let visitor = MethodVisitor {};
-            return visitor.visit_str(&val).map(Some);
-        }
-
-        Ok(None)
-    }
-
-    pub fn serialize_optional<S: Serializer>(
-        method: &Option<Method>,
-        ser: S,
-    ) -> Result<S::Ok, S::Error> {
-        if let Some(method) = method {
-            return serialize(method, ser);
-        }
-
-        ser.serialize_none()
-    }
-}
-
-// Jan 2, 2006 3:04:05 PM
-const CODEBUILD_TIME_FORMAT: &str = "%b %e, %Y %l:%M:%S %p";
-
-pub mod codebuild_time {
-    use super::*;
-    struct TimeVisitor;
-    impl<'de> Visitor<'de> for TimeVisitor {
-        type Value = DateTime<Utc>;
-
-        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-            write!(formatter, "valid codebuild time: {}", CODEBUILD_TIME_FORMAT)
-        }
-
-        fn visit_str<E: DeError>(self, val: &str) -> Result<Self::Value, E> {
-            Utc.datetime_from_str(val, CODEBUILD_TIME_FORMAT)
-                .map_err(|e| DeError::custom(format!("Parse error {} for {}", e, val)))
-        }
-    }
-
-    pub mod str_time {
-        use super::*;
-
-        pub(crate) fn deserialize<'de, D>(d: D) -> Result<DateTime<Utc>, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            d.deserialize_str(TimeVisitor)
-        }
-
-        pub fn serialize<S: Serializer>(date: &DateTime<Utc>, ser: S) -> Result<S::Ok, S::Error> {
-            let s = format!("{}", date.format(CODEBUILD_TIME_FORMAT));
-            ser.serialize_str(&s)
-        }
-    }
-
-    pub mod optional_time {
-        use super::*;
-
-        pub(crate) fn deserialize<'de, D>(
-            deserializer: D,
-        ) -> Result<Option<DateTime<Utc>>, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            let s: Option<String> = Option::deserialize(deserializer)?;
-            if let Some(val) = s {
-                let visitor = TimeVisitor {};
-                return visitor.visit_str(&val).map(Some);
-            }
-
-            Ok(None)
-        }
-
-        pub fn serialize<S: Serializer>(
-            date: &Option<DateTime<Utc>>,
-            ser: S,
-        ) -> Result<S::Ok, S::Error> {
-            if let Some(date) = date {
-                return str_time::serialize(date, ser);
-            }
-
-            ser.serialize_none()
-        }
-    }
 }
 
 #[cfg(test)]
@@ -581,125 +431,5 @@ mod test {
         };
         let encoded = serde_json::to_string(&instance).unwrap();
         assert_eq!(encoded, String::from(r#"{"v":36}"#));
-    }
-
-    #[test]
-    fn test_deserialize_missing_http_headers() {
-        #[derive(Deserialize)]
-        struct Test {
-            #[serde(deserialize_with = "http_serde::header_map::deserialize", default)]
-            pub headers: HeaderMap,
-        }
-        let data = json!({
-            "not_headers": {}
-        });
-
-        let expected = HeaderMap::new();
-
-        let decoded: Test = serde_json::from_value(data).unwrap();
-        assert_eq!(expected, decoded.headers);
-    }
-
-    type TestTime = DateTime<Utc>;
-
-    #[test]
-    fn test_deserialize_codebuild_time() {
-        #[derive(Deserialize)]
-        struct Test {
-            #[serde(with = "codebuild_time::str_time")]
-            pub date: TestTime,
-        }
-        let data = json!({
-            "date": "Sep 1, 2017 4:12:29 PM"
-        });
-
-        let expected = Utc
-            .datetime_from_str("Sep 1, 2017 4:12:29 PM", CODEBUILD_TIME_FORMAT)
-            .unwrap();
-        let decoded: Test = serde_json::from_value(data).unwrap();
-        assert_eq!(expected, decoded.date);
-    }
-
-    #[test]
-    fn test_deserialize_codebuild_optional_time() {
-        #[derive(Deserialize)]
-        struct Test {
-            #[serde(with = "codebuild_time::optional_time")]
-            pub date: Option<TestTime>,
-        }
-        let data = json!({
-            "date": "Sep 1, 2017 4:12:29 PM"
-        });
-
-        let expected = Utc
-            .datetime_from_str("Sep 1, 2017 4:12:29 PM", CODEBUILD_TIME_FORMAT)
-            .unwrap();
-        let decoded: Test = serde_json::from_value(data).unwrap();
-        assert_eq!(Some(expected), decoded.date);
-    }
-
-    #[test]
-    fn test_http_method_serializer() {
-        #[derive(Deserialize, Serialize)]
-        struct Test {
-            #[serde(with = "http_method")]
-            pub method: http::Method,
-        }
-        let data = json!({
-            "method": "DELETE"
-        });
-        let decoded: Test = serde_json::from_value(data.clone()).unwrap();
-        assert_eq!(http::Method::DELETE, decoded.method);
-
-        let recoded = serde_json::to_value(decoded).unwrap();
-        assert_eq!(data, recoded);
-    }
-
-    #[test]
-    fn test_http_optional_method_serializer() {
-        #[derive(Deserialize, Serialize)]
-        struct Test {
-            #[serde(deserialize_with = "http_method::deserialize_optional")]
-            #[serde(serialize_with = "http_method::serialize_optional")]
-            #[serde(default)]
-            pub method: Option<http::Method>,
-        }
-        let data = json!({
-            "method": "DELETE"
-        });
-        let decoded: Test = serde_json::from_value(data.clone()).unwrap();
-        assert_eq!(Some(http::Method::DELETE), decoded.method);
-
-        let recoded = serde_json::to_value(decoded).unwrap();
-        assert_eq!(data, recoded);
-
-        let data = json!({ "method": null });
-        let decoded: Test = serde_json::from_value(data.clone()).unwrap();
-        assert_eq!(None, decoded.method);
-
-        let data = json!({});
-        let decoded: Test = serde_json::from_value(data.clone()).unwrap();
-        assert_eq!(None, decoded.method);
-    }
-
-    #[test]
-    fn test_serialize_headers() {
-        #[derive(Deserialize, Serialize)]
-        struct Test {
-            #[serde(deserialize_with = "http_serde::header_map::deserialize", default)]
-            #[serde(serialize_with = "serialize_multi_value_headers")]
-            headers: HeaderMap,
-        }
-        let data = json!({
-            "headers": {
-                "Accept": ["*/*"]
-            }
-        });
-        let decoded: Test = serde_json::from_value(data).unwrap();
-        assert_eq!(&"*/*", decoded.headers.get("Accept").unwrap());
-
-        let recoded = serde_json::to_value(decoded).unwrap();
-        let decoded: Test = serde_json::from_value(recoded).unwrap();
-        assert_eq!(&"*/*", decoded.headers.get("Accept").unwrap());
     }
 }
